Lab_10/requirements.txt
Flask==3.0.0
PyJWT==2.8.0
httpx==0.25.2
Werkzeug==3.0.1
